Update Object3D.js
Add a new member variable called "viewable" that is used to notify user if the object is in the camera frustum planes. The variable is set automatically in the WebGLRenderer frustum culling process via the isObjectViewable function. The frustum culling is an expensive process, this new variable save us time in redoing the same process twice (specially when it is needed for trivial code like labels)